backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
google-genai>=1.0.0
google-cloud-firestore>=2.14.0
playwright>=1.40.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
